% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{as_posixct}
\alias{as_posixct}
\title{Turn an integer into a posixct type}
\usage{
as_posixct(x)
}
\arguments{
\item{x}{the integer to turn into a posixct}
}
\description{
Turn an integer into a posixct type
}
