% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{bt_buy}
\alias{bt_buy}
\title{Buy a cryptocurrency}
\usage{
bt_buy(market, quantity, price, dry_run = TRUE)
}
\arguments{
\item{market}{which currency to buy}

\item{quantity}{how many of the quote currency should be bought?}

\item{price}{how much (in the base currency) would you like to pay?}

\item{dry_run}{should this be run on the exchange? (default FALSE)}
}
\description{
Buy a cryptocurrency
}
