% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{bt_sell}
\alias{bt_sell}
\title{Sell a cryptocurrency}
\usage{
bt_sell(market, quantity, price, dry_run = TRUE)
}
\arguments{
\item{market}{which currency to buy}

\item{quantity}{how many of the quote currency should be sold?}

\item{price}{how much you'd like someone else to pay (in the base currency) 
for each unit the quote currency?}

\item{dry_run}{should this be run on the exchange? (default FALSE)}
}
\description{
Sell a cryptocurrency
}
