% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{cancel_open_buys}
\alias{cancel_open_buys}
\title{Cancel the open buy orders}
\usage{
cancel_open_buys(dry_run = TRUE, log_cb)
}
\arguments{
\item{dry_run}{is this a dry run (not actually run)? (default: TRUE)}

\item{log_cb}{a callback is given market, time_stamp, order_type, quantity
price, and order_uuid.}
}
\description{
Cancel the open buy orders
}
