% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{cancel_open_orders}
\alias{cancel_open_orders}
\title{Cancel all open orders}
\usage{
cancel_open_orders(dry_run = FALSE)
}
\arguments{
\item{dry_run}{should this be run on the exchange? (default FALSE)}
}
\description{
Cancel all open orders
}
