% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{execute_buy}
\alias{execute_buy}
\title{Place a buy order}
\usage{
execute_buy(market, quantity, rate, dry_run = TRUE, log_cb = NULL)
}
\arguments{
\item{market}{the market to place the buy limit order on.}

\item{quantity}{the quantity of the transaction currency to buy.}

\item{rate}{the price you are willing to pay per unit of the transaction
currency.}

\item{dry_run}{is this a dry run (not actually run)? (default: TRUE)}

\item{log_cb}{a callback is given market, time_stamp, order_type, quantity
price, and order_uuid. (default: NULL)}
}
\description{
Place a buy order
}
