% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{execute_sell}
\alias{execute_sell}
\title{Place a sell order}
\usage{
execute_sell(market, quantity, rate, dry_run = TRUE, log_cb = NULL)
}
\arguments{
\item{market}{the market to place the buy limit order on.}

\item{quantity}{the quantity of the base currency to sell.}

\item{rate}{the price you are willing to pay per unit of the quote
currency.}

\item{dry_run}{is this a dry run (not actually run)? (default: TRUE)}

\item{log_cb}{a callback is given market, time_stamp, order_type, quantity
price, and order_uuid. (default: NULL)}
}
\description{
Place a sell order
}
