% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{flip_market_name}
\alias{flip_market_name}
\title{Flip the base and quote currency for a market}
\usage{
flip_market_name(x)
}
\arguments{
\item{x}{the vector of markets to flip.}
}
\description{
Flip the base and quote currency for a market
}
