% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{flip_market_summary}
\alias{flip_market_summary}
\title{Flip Bittrex market summaries}
\usage{
flip_market_summary(x)
}
\arguments{
\item{x}{the market summary}
}
\description{
Flip Bittrex market summaries
}
