% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{flip_markets}
\alias{flip_markets}
\title{Flip a market}
\usage{
flip_markets(market_name, high, low, volume, last, base_volume, time_stamp, bid,
  ask, open_buy_orders, open_sell_orders, prev_day, created)
}
\arguments{
\item{market_name}{the market names.}

\item{high}{the vector of highs.}

\item{low}{the vector of lows.}

\item{volume}{the vector of volumes.}

\item{last}{the vector of last prices.}

\item{base_volume}{the vector of base_volumes - this gets set to NA.}

\item{time_stamp}{the vector of time stamps.}

\item{bid}{the vecor of bids.}

\item{ask}{the vecor of asks.}

\item{open_buy_orders}{the vector of open buy orders.}

\item{open_sell_orders}{the vector of open sell orders.}

\item{prev_day}{the previous day's close.}

\item{created}{the vector of dates the market was created.}
}
\description{
Flip a market
}
