% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{get_result}
\alias{get_result}
\title{If a bittrex call was successful, return the data}
\usage{
get_result(x)
}
\arguments{
\item{x}{the return of a call to the Bittrex exchange}
}
\description{
If a bittrex call was successful, return the data
}
